pyproject.toml
[project]
name = "speechrecognitionproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "better-profanity>=0.7.0",
    "openai-whisper>=20240930",
    "pyannote-audio>=3.3.2",
    "pyaudio>=0.2.14",
    "pydantic>=2.10.6",
    "pydub>=0.25.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "speechrecognition>=3.14.1",
    "textblob>=0.19.0",
]
